feat: Update dependencies and refine migration scripts
This commit updates the `hono` and `pothos-drizzle-generator` dependencies to their latest versions. It also refines the `drizzle:reset` script to include `drizzle:migrate` and `drizzle:seed` for a more complete reset process. Additionally, the `DATABASE_URL` in `.env` is simplified by removing the `?schema=sample` suffix. The initialization steps in `README.md` have been updated to reflect the changes in the `package.json` scripts.